Move `pages` into src, convert another file
